7r3A Build 10 advance 	merged 64-bit / 32-bit sources 	nothing is called long and all c language i32s are called int 	correction to 1st / 2nd pass behaviour of $text translations

+++ masmx.7r3/src/iqltable.c
@@ -23,7 +23,7 @@
 
 
 
-static object *insert_qltable(char *l, long equator, int type)
+static object *insert_qltable(char *l, int equator, int type)
 {
    line_item		 local = zero_o;
